Introduce provider abstraction and reference implementations for LLM + OCR (issue #615)
