+++ app/ai-service/services/humanitarian_verification.py
@@ -6,12 +6,10 @@ from typing import Any, Dict, List, Optional
 import time
 import metrics
 
-import httpx
-
 from config import settings
 from services.humanitarian_prompt import HumanitarianPromptEngine
 from services.circuit_breaker import CircuitBreaker
-from services.test_provider import TestProvider
+from services.provider_factory import get_llm_provider
 from exceptions import AIServiceError
 
 logger = logging.getLogger(__name__)
@@ -22,7 +20,7 @@ class HumanitarianVerificationService:
 
     def __init__(self):
         self.prompt_engine = HumanitarianPromptEngine()
-        self.test_provider = TestProvider()
+        self.llm_provider = get_llm_provider()
         self.breakers = {
             "openai": CircuitBreaker(
                 name="openai",
@@ -144,133 +142,13 @@ class HumanitarianVerificationService:
         user_prompt: str,
         timeout: Optional[float] = None,
     ) -> str:
-        if provider == "test":
-            return self._call_test(model, system_prompt, user_prompt)
-        if provider == "openai":
-            return self._call_openai(model, system_prompt, user_prompt, timeout)
-        if provider == "groq":
-            return self._call_groq(model, system_prompt, user_prompt, timeout)
-        raise ValueError(f"Unsupported provider: {provider}")
-
-    def _call_openai(
-        self,
-        model: str,
-        system_prompt: str,
-        user_prompt: str,
-        timeout: Optional[float] = None,
-    ) -> str:
-        if not settings.openai_api_key:
-            raise RuntimeError("OpenAI API key is not configured")
-        return self._call_chat_completion_api(
-            base_url="https://api.openai.com/v1/chat/completions",
-            api_key=settings.openai_api_key,
+        return self.llm_provider.send_chat_completion(
+            provider=provider,
             model=model,
             system_prompt=system_prompt,
             user_prompt=user_prompt,
             timeout=timeout,
         )
-
-    def _call_groq(
-        self,
-        model: str,
-        system_prompt: str,
-        user_prompt: str,
-        timeout: Optional[float] = None,
-    ) -> str:
-        if not settings.groq_api_key:
-            raise RuntimeError("Groq API key is not configured")
-        return self._call_chat_completion_api(
-            base_url="https://api.groq.com/openai/v1/chat/completions",
-            api_key=settings.groq_api_key,
-            model=model,
-            system_prompt=system_prompt,
-            user_prompt=user_prompt,
-            timeout=timeout,
-        )
-
-    def _call_chat_completion_api(
-        self,
-        base_url: str,
-        api_key: str,
-        model: str,
-        system_prompt: str,
-        user_prompt: str,
-        timeout: Optional[float] = None,
-    ) -> str:
-        if settings.ai_deterministic_mode:
-            logger.info("Deterministic AI mode enabled: returning stable response")
-            return self._get_deterministic_response(model, system_prompt, user_prompt)
-
-        payload = {
-            "model": model,
-            "temperature": 0.1,
-            "messages": [
-                {"role": "system", "content": system_prompt},
-                {"role": "user", "content": user_prompt},
-            ],
-        }
-        headers = {
-            "Authorization": f"Bearer {api_key}",
-            "Content-Type": "application/json",
-        }
-
-        req_timeout = timeout if timeout is not None else float(settings.llm_timeout_seconds)
-        provider_name = "openai" if "openai" in base_url else "groq"
-
-        try:
-            with httpx.Client(timeout=req_timeout) as client:
-                response = client.post(base_url, json=payload, headers=headers)
-                response.raise_for_status()
-                data = response.json()
-        except httpx.TimeoutException as exc:
-            logger.error("LLM provider %s request timed out after %s seconds", provider_name, req_timeout)
-            raise AIServiceError(
-                message=f"LLM request timed out after {req_timeout}s",
-                code="AI_TIMEOUT",
-                details={"provider": provider_name, "timeout_seconds": req_timeout},
-            ) from exc
-        except httpx.HTTPStatusError as exc:
-            logger.error("LLM provider %s returned status %s: %s", provider_name, exc.response.status_code, exc.response.text)
-            raise AIServiceError(
-                message=f"LLM request failed with status {exc.response.status_code}",
-                code="AI_PROVIDER_ERROR",
-                details={"provider": provider_name, "status_code": exc.response.status_code},
-            ) from exc
-        except Exception as exc:
-            logger.error("LLM provider %s connection or unexpected error: %s", provider_name, str(exc))
-            raise AIServiceError(
-                message=f"LLM connection error: {str(exc)}",
-                code="AI_CONNECTION_ERROR",
-                details={"provider": provider_name},
-            ) from exc
-
-        try:
-            content = data["choices"][0]["message"]["content"]
-        except (KeyError, IndexError, TypeError) as exc:
-            raise RuntimeError(f"Unexpected LLM response format: {data}") from exc
-
-        if not content:
-            raise RuntimeError("LLM returned empty content")
-
-        return str(content)
-
-    def _call_test(self, model: str, system_prompt: str, user_prompt: str) -> str:
-        response = self.test_provider.get_response(
-            endpoint="humanitarian",
-            request_data={
-                "system_prompt": system_prompt,
-                "user_prompt": user_prompt,
-            },
-        )
-        return json.dumps(response, separators=(",", ":"), sort_keys=True)
-
-    def _get_deterministic_response(self, model: str, system_prompt: str, user_prompt: str) -> str:
-        stable_response = {
-            "verdict": "credible",
-            "confidence": 0.74,
-            "summary": "Deterministic verification output for testing",
-        }
-        return json.dumps(stable_response, separators=(",", ":"), sort_keys=True)
 
     def _parse_json_response(self, content: str) -> Dict[str, Any]:
         normalized = content.strip()
