+++ app/ai-service/services/provider_types.py
@@ -1,0 +1,14 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class FieldMatch:
+    value: str
+    confidence: float
+
+
+@dataclass
+class OCRResult:
+    fields: dict[str, FieldMatch]
+    raw_text: str
+    processing_time_ms: int
